refactor (AW-303): split detail components so data can be passed in
