+++ nextjs/src/app/[locale]/(pages)/[...slug]/page.tsx
@@ -1,15 +1,9 @@
 import { getPage } from "@/lib/strapi"
-import { NavProvider } from "@/components/nav-bar/nav-context"
-import NavBar from "@/components/nav-bar/nav-bar"
-import HeroWrapper from "@/components/stapi/hero-wrapper"
-import Intro from "@/components/intro"
-import Text from "@/components/text"
-import { renderSections } from "@/components/dynamic-zone/render-sections"
-import Footer from "@/components/stapi/footer"
 import { Locale } from "@/lib/locale"
 import { Metadata } from "next"
 import { ABSOLUTE_URL } from "@/lib/absolute-url"
 import { buildMetadata } from "@/lib/metadata"
+import PageDetail from "@/app/[locale]/(pages)/page-detail"
 
 export async function generateMetadata(props: {
   params: Promise<{
@@ -57,36 +51,5 @@ export default async function LandingPage(props: {
     isActive: true,
   }
   const data = await getPage(locale, URI_PATH)
-  const locales = data.localizations.map(
-    (item: { locale: Locale; slug: string }) => {
-      return {
-        href: `/${item.locale.toLowerCase()}/${item.slug}`,
-        locale: item.locale,
-        isActive: false,
-      }
-    },
-  )
-  locales.push(activeLocale)
-
-  const { hero, intro, sections } = data
-
-  return (
-    <>
-      <NavProvider>
-        <NavBar items={locales} />
-        {hero && <HeroWrapper hero={hero} />}
-      </NavProvider>
-      {intro && (
-        <Intro>
-          <Text markdown={intro} />
-        </Intro>
-      )}
-      {sections &&
-        sections.length > 0 &&
-        sections.map((section: any, index: number) =>
-          renderSections(section, index, activeLocale.locale),
-        )}
-      <Footer locale={activeLocale.locale} />
-    </>
-  )
+  return <PageDetail data={data} activeLocale={activeLocale} />
 }
