+++ nextjs/src/app/[locale]/blogs/[slug]/page.tsx
@@ -1,16 +1,9 @@
-import { getLocaleDateFormatted, Locale } from "@/lib/locale"
+import { Locale } from "@/lib/locale"
 import { Metadata } from "next"
 import { getBlogPage } from "@/lib/strapi"
 import { BLOG_SLUGS } from "@/lib/slugs"
 import { buildMetadata } from "@/lib/metadata"
-import { NavProvider } from "@/components/nav-bar/nav-context"
-import NavBar from "@/components/nav-bar/nav-bar"
-import HeroWrapper from "@/components/stapi/hero-wrapper"
-import InfoLabel from "@/components/info-label"
-import TextImage from "@/components/text-image"
-import { renderSections } from "@/components/dynamic-zone/render-sections"
-import Footer from "@/components/stapi/footer"
-import { getDictionary } from "@/lib/get-dictionary.server"
+import BlogDetail from "@/app/[locale]/blogs/blog-detail"
 
 export async function generateMetadata(props: {
   params: Promise<{
@@ -49,42 +42,5 @@ export default async function BlogPage(props: {
     isActive: true,
   }
   const data = await getBlogPage(activeLocale.locale, slug)
-  const { hero, main_blog, sections, publishedAt, createdAt } = data
-  const locales = data.localizations.map(
-    (item: { locale: Locale; slug: string }) => {
-      return {
-        href: `/${item.locale.toLowerCase()}/${BLOG_SLUGS[item.locale.toLowerCase() as Locale]}/${item.slug}`,
-        locale: item.locale,
-        isActive: false,
-      }
-    },
-  )
-  locales.push(activeLocale)
-  const dictionary = await getDictionary(activeLocale.locale as Locale)
-
-  return (
-    <>
-      <NavProvider>
-        <NavBar items={locales} />
-        {hero && <HeroWrapper hero={hero} />}
-      </NavProvider>
-      <section className={"px-4 lg:px-0 relative bg-white"}>
-        <div className="container sm:px-2">
-          <div className="mx-auto pb-8 max-w-4xl">
-            <InfoLabel
-              text={`${dictionary.common.publishedAt} ${getLocaleDateFormatted({ date: publishedAt ?? createdAt, locale: activeLocale.locale as Locale })}`}
-              className="block mb-4"
-            />
-            <TextImage markdown={main_blog} />
-          </div>
-        </div>
-      </section>
-      {sections &&
-        sections.length > 0 &&
-        sections.map((section: any, index: number) =>
-          renderSections(section, index, activeLocale.locale),
-        )}
-      <Footer locale={activeLocale.locale} />
-    </>
-  )
+  return <BlogDetail data={data} activeLocale={activeLocale} />
 }
